Namespacing Speech Services
