+++ packages/component/src/SpeechServices/SpeechSDK.js
