+++ packages/component/src/SpeechServices/SpeechToText.js
