+++ packages/component/src/SpeechServices/SpeechToText/createSpeechRecognitionPonyfill.test.js
