+++ packages/component/src/SpeechServices/TextToSpeech/SpeechSynthesisVoice.js
