+++ packages/component/src/SpeechServices/TextToSpeech/fetchSpeechData.js
