+++ packages/component/src/SpeechServices/TextToSpeech/subscribeEvent.js
