+++ packages/component/src/SpeechServices/fetchAuthorizationToken.js
