+++ packages/component/src/index.js
@@ -1,16 +1,3 @@
-import createSpeechRecognitionPonyfill from './SpeechToText';
-import createSpeechSynthesisPonyfill from './TextToSpeech';
-import fetchAuthorizationToken from './fetchAuthorizationToken';
+import SpeechServices from './SpeechServices';
 
-export default async function (...args) {
-  return {
-    ...(await createSpeechRecognitionPonyfill(...args)),
-    ...(await createSpeechSynthesisPonyfill(...args))
-  };
-};
-
-export {
-  createSpeechRecognitionPonyfill,
-  createSpeechSynthesisPonyfill,
-  fetchAuthorizationToken
-}
+export default SpeechServices
